Switched from using build_design_matrices to a compiled eval function.
build_design_matrices will compile an eval for each factor or function
of a factor, but it does it each time build_design_matrices is called.
Since we want to expand the design point a large number of times in
coordinate exchange it's much faster to compile an expression that
generates a point once, then eval it multiple times in the loops.